requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
google-api-core==1.17.0
google-auth==1.14.1
google-cloud==0.34.0
google-cloud-vision==1.0.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
idna==2.9
numpy==1.18.3
opencv-contrib-python==3.4.3.18
opencv-python==4.1.1.26
picamera==1.13
pkg-resources==0.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pytz==2020.1
requests==2.23.0
RPi.GPIO==0.7.0
rsa==4.0
six==1.14.0
twilio==6.39.0
urllib3==1.25.9
